@@ -1,105 +1,69 @@
 <template>
-  <div id="app">
-    <transition name="fade"
-        mode="out-in">
-      <router-view></router-view>
+  <RouterView v-slot="{ Component }">
+    <transition name="fade-right" mode="out-in">
+      <component :is="Component" />
     </transition>
-  </div>
+  </RouterView>
 </template>
-<script>
-  export default {
-    name: 'app'
-  }
+
+<script setup>
+import { RouterView } from 'vue-router'
 </script>
 
 <style lang="scss">
-  body {
-    margin: 0px;
-    padding: 0px;
-    /*background: url(assets/bg1.jpg) center !important;
-		background-size: cover;*/
-    // background: #1F2D3D;
-    font-family: Helvetica Neue, Helvetica, PingFang SC, Hiragino Sans GB, Microsoft YaHei, SimSun, sans-serif;
-    font-size: 14px;
-    -webkit-font-smoothing: antialiased;
-  }
+.card-header{
+  display: flex;
+  justify-content: right;
+}
+.card-header2{
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+.aside-menu-icon{
+  width: 25px;
+  height: 25px;
+  margin-right: 10px;
+}
+body {
+  margin: 0;
+  padding: 0;
+  font-family: Helvetica Neue, Helvetica, PingFang SC, Hiragino Sans GB, Microsoft YaHei, SimSun, sans-serif;
+  font-size: 14px;
+  -webkit-font-smoothing: antialiased;
+}
 
-  #app {
-    position: absolute;
-    top: 0px;
-    bottom: 0px;
-    width: 100%;
-  }
-
-  .el-submenu [class^=fa] {
-    vertical-align: baseline;
-    margin-right: 10px;
-  }
-
-  .el-menu-item [class^=fa] {
-    vertical-align: baseline;
-    margin-right: 10px;
-  }
-
-  .toolbar {
-    background: #f2f2f2;
-    padding: 10px;
-    //border:1px solid #dfe6ec;
-    margin: 10px 0px;
-    .el-form-item {
-      margin-bottom: 10px;
-    }
-  }
-
-  .fade-enter-active,
-  .fade-leave-active {
-    transition: all .2s ease;
-  }
-
-  .fade-enter,
-  .fade-leave-active {
-    opacity: 0;
-  }
-  .clearfix:before,
-  .clearfix:after {
-    display: table;
-    content: "";
-  }
-  .clearfix:after {
-    clear: both
-  }
-
-  .avatar-uploader .el-upload {
-    border: 1px dashed #d9d9d9;
-    border-radius: 6px;
-    cursor: pointer;
-    position: relative;
-    overflow: hidden;
-  }
-  .avatar-uploader .el-upload:hover {
-    border-color: #20a0ff;
-  }
-  .el-upload {
-    position: relative;
-    overflow: hidden;
-  }
-  .el-upload:hover .panel{
-    left:0
-  }
-  .panel {
-    position: absolute;
-    box-sizing: border-box;
-    top: 50%;
-    transform: translateY(-50%);
-    padding: 0;
-    background: rgba(0, 0, 0, 0.6);
-    -webkit-border-radius: 0 6px 6px 0;
-    -moz-border-radius: 0 6px 6px 0;
-    border-radius: 0 6px 6px 0;
-    color: #fff;
-    left: -40px;
-    transition: all 0.3s;
-    height: 30px;
-    line-height: 30px;
-  }
+.avatar-uploader .el-upload {
+  border: 1px dashed #d9d9d9;
+  border-radius: 6px;
+  cursor: pointer;
+  position: relative;
+  overflow: hidden;
+}
+.avatar-uploader .el-upload:hover {
+  border-color: #20a0ff;
+}
+.el-upload {
+  position: relative;
+  overflow: hidden;
+}
+.el-upload:hover .panel{
+  left:0
+}
+.panel {
+  position: absolute;
+  box-sizing: border-box;
+  top: 50%;
+  transform: translateY(-50%);
+  padding: 0;
+  background: rgba(0, 0, 0, 0.6);
+  -webkit-border-radius: 0 6px 6px 0;
+  -moz-border-radius: 0 6px 6px 0;
+  border-radius: 0 6px 6px 0;
+  color: #fff;
+  left: -40px;
+  transition: all 0.3s;
+  height: 30px;
+  line-height: 30px;
+}
 </style>
