@@ -1,9 +1,8 @@
-let _host = null;
 const host = function(){
-	if(_host) return _host;
 	let ln = window.location;
-	//ent.cw.net
-	//http://localhost:7097
-	return ln.href.indexOf("localhost") > 0 ? 'http://localhost' : "https://ent.uqiantu.com";
+	console.log(ln.href.indexOf("localhost"))
+	return ln.href.indexOf("localhost") > 0 ? 'http://localhost:5173' : "https://ent.uqiantu.com";
 };
-global.host = host();
+export default {
+	baseURL: host()
+};
